整理代码，减少SDK修改
