+++ MPChartExample/src/main/res/layout/activity_linechart.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.github.mikephil.charting.charts.LineChart
+    <com.xxmassdeveloper.mpchartexample.custom.LineChartCustom
         android:id="@+id/chart1"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
